updated variational math and way to nest pages 1 deep
**note for all** before this commit, pages were sorted by name. In
order to nest pages into a tree format, they needed to be sorted by
some other more easily adjustable value. If your page does not have an
index in the header (the part where you declare the page title, etc.)
it will appear before the pages with an index, and in no particular
order. If you’d like to add an index, just add the line `index: 1`
after you give the layout and the title. If you want to nest pages, you
need to change the layout of the subpages to `layout: section`.